Remove unnecessary build commands, split sample application into different classes
